requirements.txt

bcrypt==4.1.1
blinker==1.7.0
certifi==2023.11.17
gitdb==4.0.11
GitPython==3.1.40
matplotlib==3.8.2
numpy==1.26.2
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
pyarrow==14.0.1
pydeck==0.8.1b0
Pygments==2.17.2
pymongo==4.6.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
referencing==0.32.0
requests==2.31.0
rich==13.7.0
rpds-py==0.13.2
six==1.16.0
smmap==5.0.1
streamlit==1.29.0
validators==0.22.0

